Implement HuffmanTree
HuffmanCode is so purpular compressing algorithm
It is made by frequence and minHeap
I Implemented last year in Data structure. but I did not implement
properly. I did not implement minheap. so I use priority queue in library.
so this code have been lagacy.
I try to implement by c but It is hard to processing string. so I
implement by java.